Revert "Remove Orphan Instance of Foldable"
This reverts commit ae6083ad61ea4f0be7523c476266d677a68de4d7.

Turns out that base-orphans is required to get the instance of Foldable
we want. If there's going to be orphan instances no matter what, it
might make sense to just stick with our manual instance of Foldable
instead of pulling in a whole lot of others.

In the meantime, going to appease Travis with this revert.